requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.0.1
numpy==1.26.4
Pillow==9.5.0
tensorflow==2.14.0
gunicorn==20.1.0
